Resource - Add Ingredient API ADD/UPDATE/DELETE end points

+++ webapp/resource/impl/src/main/java/cmahy/webapp/resource/impl/adapter/taco/shop/repository/ClientOrderRepositoryImpl.java
@@ -20,7 +20,7 @@ public interface ClientOrderRepositoryImpl extends
 
     @Override
     default ClientOrderPage getByUser(User user, EntityPageable pageable) {
-        Page<ClientOrder> all = ClientOrderRepositoryImpl.this.findByUser(user, PageRequest.of(pageable.pageNumber().intValue(), pageable.pageSize()));
+        Page<ClientOrder> all = ClientOrderRepositoryImpl.this.findByUser(user, PageRequest.of(pageable.pageNumber(), pageable.pageSize()));
 
         return new ClientOrderPage(
             all.getContent(),
